backend/requirements.txt
fastapi
uvicorn
langchain
langchain-ollama
python-dotenv